refactor: address fourth review feedback — cleanup and DRY
- Extract escapeXml() to src/utils/xml.ts, imported by both
  just-bash-agent-environment.ts and the new skill-invocation.ts
- Extract buildSkillInvocationXml() to src/session/skill-invocation.ts
  so it is internal to the package but directly importable by tests
- Remove unused AgentMessage and ExtensionContext imports from
  skill-commands.test.ts
- Fix misleading "synchronously" comment in addSkill() to accurately
  describe the fire-and-forget async write
