+++ packages/otter-agent/src/session/agent-session.ts
@@ -18,12 +18,12 @@ import type { Extension } from "../extensions/extension.js";
 import type { AgentEnvironment } from "../interfaces/agent-environment.js";
 import type { AuthStorage } from "../interfaces/auth-storage.js";
 import type { EntryId, SessionManager } from "../interfaces/session-manager.js";
-import type { SkillDefinition } from "../interfaces/skill-definition.js";
 import { isSkillSupportedAgentEnvironment } from "../interfaces/skill-supported-agent-environment.js";
 import type { ToolDefinition } from "../interfaces/tool-definition.js";
 import type { UIProvider } from "../interfaces/ui-provider.js";
 import { convertToLlm } from "./messages.js";
 import { ModelRegistry } from "./model-registry.js";
+import { buildSkillInvocationXml } from "./skill-invocation.js";
 import { buildSystemPrompt } from "./system-prompt.js";
 import { wrapToolDefinition } from "./tool-wrapper.js";
 
@@ -696,51 +696,4 @@ export class AgentSession {
 			this._extensionRunner.registerCommand(`skill:${skill.name}`, commandOptions);
 		}
 	}
-}
-
-// ─── Helpers ─────────────────────────────────────────────────────────────────
-
-/** Escape characters that are special in XML/HTML attributes and text content. */
-function escapeXml(str: string): string {
-	return str
-		.replace(/&/g, "&amp;")
-		.replace(/</g, "&lt;")
-		.replace(/>/g, "&gt;")
-		.replace(/"/g, "&quot;")
-		.replace(/'/g, "&apos;");
-}
-
-/**
- * Build the XML block injected into the conversation when a skill is invoked.
- * Follows the same format as pi-coding-agent's skill invocation messages.
- *
- * @param skill - The skill definition to invoke.
- * @param args - Arguments passed by the user after the command name.
- * @param filePath - Optional absolute path to the skill's SKILL.md file.
- *   When provided, a `location` attribute and relative-path hint are included
- *   so the agent can resolve supporting files referenced in the skill content.
- *   Omit for self-contained skills or when the path is not available.
- */
-export function buildSkillInvocationXml(
-	skill: SkillDefinition,
-	args: string,
-	filePath?: string,
-): string {
-	const parts: string[] = [];
-
-	if (filePath) {
-		const skillDir = filePath.substring(0, filePath.lastIndexOf("/"));
-		parts.push(`<skill name="${escapeXml(skill.name)}" location="${escapeXml(filePath)}">`);
-		parts.push(`References are relative to ${skillDir}/.`);
-	} else {
-		parts.push(`<skill name="${escapeXml(skill.name)}">`);
-	}
-
-	parts.push("", skill.content, "</skill>");
-
-	const trimmedArgs = args.trim();
-	if (trimmedArgs) {
-		parts.push("", trimmedArgs);
-	}
-	return parts.join("\n");
 }
